Разобьет utils.js на три отдельных файла
